refactor(protocols,http): split protocol models and gateway test fixtures
Break the mock protocol implementations and MCP schema types into focused modules so maintainers and agents can navigate the codebase by responsibility instead of large mixed files. Reorganize gateway tests around shared fixtures and behavior domains, remove stale segment files, and document the new layout while keeping the full CI pipeline green.
